requirements.txt
langchain
streamlit
langchain-community
langchain-google-genai
python-dotenv
